requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
passlib==1.7.4
bcrypt==4.2.1
python-jose==3.3.0
python-dotenv==1.0.1
python-multipart==0.0.20
